Added employee stream serializer

+++ src/main/java/com/kryptnostic/conductor/ConductorSparkImpl.java
@@ -37,12 +37,12 @@ public class ConductorSparkImpl implements ConductorSparkApi {
                 e.getAs( "title" ),
                 e.getAs( "salary" ) ) ).collect();
         highlyPaidEmps.forEach( e -> logger.info( e ) );
-        
+
         return Lists.newArrayList( emps.javaRDD().map( e -> new Employee(
                 e.getAs( "name" ),
                 e.getAs( "dept" ),
                 e.getAs( "title" ),
-                e.getAs( "salary" ) ) ).collect() );
+                (int) e.getAs( "salary" ) ) ).collect() );
     }
 
 }
